Create Self-Replicating Virus
This python virus, much like the SARS-CoV-2 virus, is designed to spread from host to host and has the ability to replicate itself. In more technical terms, we will be writing a program to infect all the python files present in 6ft distance (same directory) with the self-replicating code and perform malicious activities through the infected python files.