Update advanced Trojan virus added a new feature
Update Version 1.1:
includes a new feature that enables the system to speak.
Update Version 1.2:
includes a new feature that downloads python 3.9 from the web and runs it for u but the user needs to install the setup manually using the setup that the code downloaded.
Update Version 1.3:
includes a small rewrite of the code to enable future updates for future features and small bugs fixes.
update Version 1.4:
includes a major bug fix that didn't allow the rest of the code pass through or simply execute 
and fix some grammar issues.
update Version 1.5:
Fix the "Please touch the exit button, this window popped up as bug" appeared very large.
eliminated the ability for the program to pop up a cmd or command prompt when doing the tk command.
update Version 1.6:
added two new features. 
The code downloads python modules and added a feature that makes the cpu run at 100% capacity.
added a ability for the system to speak this words "You just got infected LAUGH MY FUCKIN ASS OUT"
update Version 1.7:
added a feature that kills the cpu by overheating it to its maximum
update Version 1.8:
Added a feature that creates a mini self replicating virus inside a file and then it executes the file.
small bugs fixes.